backend/agriculture/requirements.txt
asgiref==3.8.1
Django==4.2.14
django-environ==0.11.2
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
